requirements.txt
smbus2
pynmea2
pyserial
